Make thunk code generation private to d-decls.cc
